cmake: Add a bunch of warnings to the debug build
Hopefully, all of these are supported on the compilers people are
using.  I didn't check, though, and I also didn't spend enough
time with cmake to make these conditional on their support in the
system's C++ compiler.  Yell if building broke somewhere.

+++ src/filetransferrepeater.cpp
@@ -391,7 +391,7 @@ void FiletransferRepeater::registerXfer(PurpleXfer *xfer) {
 void FiletransferRepeater::fileSendStart() {
 	Log("ReceiveFileStraight", "fileSendStart!" << m_from.full() << " " << m_to.full());
 	Log("ReceiveFileStraight", m_sid);
-	Log("ReceiveFileStraight", m_type);
+	Log("ReceiveFileStraight", (int)m_type);
 	
 	Transport::instance()->acceptFT(m_to, m_sid, m_type, m_from.resource().empty() ? std::string(m_from.bare() + "/bot") : m_from);
 }
